RaiseMetrics 1.0 build 4
1) The launch image has the parameter set to center image, so it’s no
longer stretched in the horizontal orientation.
2) The feature list is changed for iPad devices, so now we have a card
on the center of a screen with the size in dots of iPhone 4.7-inch (375
x 667) in both horizontal and portrait orientations. If there is a
header image available, it’s set as a screen background and the header
color sets for header itself. The background image mode is set to
aspect fill.
3) After some testings I’v figured out that parallax affect doesn’t
work fine if the parallax image with the height of a screen height and
saved proportions is shorter than the 1.5 of device’s display width. So
in this case we’ll just put an image as a background for a view
controller. This situation appears mainly for horizontal modes on
iPhone and iPad devices.
4) Added the aspect fill mode for a background image for page on the
Onboarding.